+++ app/routes/_index.tsx
@@ -1,3 +1,5 @@
+import { List, Text } from "@mantine/core";
+import { Link } from "@remix-run/react";
 import { createServerClient, parse, serialize } from "@supabase/ssr";
 import { type LoaderFunctionArgs, redirect } from "@vercel/remix";
 
@@ -26,8 +28,38 @@ export async function loader({ request }: LoaderFunctionArgs) {
     }
   );
   const user = await supabase.auth.getUser();
-  if (!user.data.user) {
-    return redirect("/login");
-  }
-  return redirect("/home");
+  if (!user.data.user) return redirect("/login");
+
+  return null;
+}
+
+export default function Home() {
+  const LINKS = [
+    {
+      href: "/record",
+      label: "Start Workout",
+    },
+    {
+      href: "/workout_menus",
+      label: "Manage workout menus",
+    },
+    {
+      href: "/progress/chart",
+      label: "View progress",
+    },
+  ];
+
+  return (
+    <div style={{ height: "100%" }}>
+      <List>
+        {LINKS.map((link) => (
+          <List.Item key={link.href}>
+            <Link to={link.href}>
+              <Text size="xl">{link.label}</Text>
+            </Link>
+          </List.Item>
+        ))}
+      </List>
+    </div>
+  );
 }
