+++ app/routes/workout_menus.$id.tsx
@@ -7,7 +7,7 @@ import {
 import { Form, useLoaderData } from "@remix-run/react";
 import { createServerClient, parse, serialize } from "@supabase/ssr";
 import type { Database } from "~/types/supabase";
-import WorkoutModal from "~/components/WorkoutModal";
+import WorkoutModal from "~/components/workoutModal";
 
 export const loader = async ({ request, params }: LoaderFunctionArgs) => {
   if (!process.env.SUPABASE_URL || !process.env.SUPABASE_ANON_KEY)
@@ -33,6 +33,8 @@ export const loader = async ({ request, params }: LoaderFunctionArgs) => {
       },
     }
   );
+  const user = await supabase.auth.getUser();
+  if (!user.data.user) return redirect("/login");
   const workoutMenuId = params.id;
   if (!workoutMenuId) {
     return redirect("/workout_menus");
